avatar::memorial
moved memorial out of player into the new avatar class
