+++ src/avatar.h
@@ -13,6 +13,9 @@ class avatar : public player
         void load( JsonObject &data );
         void serialize( JsonOut &josn ) const;
         void deserialize( JsonIn &json );
+
+        /** Prints out the player's memorial file */
+        void memorial( std::ostream &memorial_file, const std::string &epitaph );
 };
 
 #endif
